requirements.txt
Package                   Version
------------------------- ---------
annotated-types           0.7.0
anthropic                 0.60.0
anyio                     4.9.0
attrs                     25.3.0
certifi                   2025.7.14
chardet                   5.2.0
charset-normalizer        3.4.2
click                     8.2.1
distro                    1.9.0
exceptiongroup            1.3.0
h11                       0.16.0
httpcore                  1.0.9
httpx                     0.28.1
idna                      3.10
jiter                     0.10.0
jsonref                   1.1.0
jsonschema                4.25.0
jsonschema-path           0.3.4
jsonschema-specifications 2025.4.1
lazy-object-proxy         1.11.0
openai                    1.98.0
openapi-schema-validator  0.6.3
openapi-spec-validator    0.7.2
packaging                 25.0
pathable                  0.4.4
pathlib                   1.0.1
pip                       22.0.2
prance                    25.4.8.0
pydantic                  2.11.7
pydantic_core             2.33.2
python-dotenv             1.1.1
PyYAML                    6.0.2
referencing               0.36.2
requests                  2.32.4
rfc3339-validator         0.1.4
rpds-py                   0.26.0
ruamel.yaml               0.18.14
ruamel.yaml.clib          0.2.12
setuptools                59.6.0
six                       1.17.0
sniffio                   1.3.1
tqdm                      4.67.1
typing_extensions         4.14.1
typing-inspection         0.4.1
urllib3                   2.5.0
